Rename descendants to subtasks and fix subtask dates
Renamed the descendants value because the option is called subtasks as
well, and descendants is actually incorrect as the copy function works
with children, not descendants.

Fixes subtask dates to have the correct start & due date, and save the
parent task a second time after setting the subtasks to force the proper
date. Because changing subtasks dates change the parent task dates.